refactor(admin): restructure admin folder and move user model to new user app
- Updated the folder structure of the "admin" app for better organization.
- Removed the "user" model from the "admin" app.
- Created a new "user" app to manage all user-related functionality separately from the admin.
