+++ apps/users/app.py
@@ -1,6 +1,6 @@
 """
 
-Authentication app module
+Common app module
 
 This file is subject to the terms and conditions defined in file 'LICENSE',
 which is part of this source code package.
@@ -11,4 +11,3 @@ from fastapi import FastAPI
 
 def register(app: FastAPI):
     """Register app instance."""
-    # Include routers
