+++ fastapi__template/admin/methods/func_util_methods.py
@@ -2,8 +2,8 @@ import sqlalchemy as sa
 from fastapi import FastAPI
 from sqladmin import Admin
 
-from apps.admin.admin import UserAdmin
 from apps.authentication.admin import ApiKeyAdmin
+from apps.users.admin import UserAdmin
 from fastapi__template.settings import SETTINGS
 
 engine = sa.create_engine(SETTINGS.DATABASE_ENGINE, poolclass=sa.StaticPool)
