+++ fastapi__template/app.py
@@ -10,7 +10,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi_sqlalchemy import DBSessionMiddleware
 
-from apps.admin.methods.func_util_methods import register_admin_views
+from fastapi__template.admin.methods.func_util_methods import register_admin_views
 from fastapi__template.dependencies import find_models, install_apps
 from fastapi__template.settings import SETTINGS
